Monkey-patch simulation
Modularize compute_aggregate, datafrale_by_entity to simulation, compute_pivot_table to Simulation, new_simulation

Add quantile computation
