+++ openfisca_survey_manager/google_colab.py
@@ -1,5 +1,3 @@
-
-
 import configparser
 import os
 
@@ -8,7 +6,7 @@ from openfisca_survey_manager import default_config_files_directory as config_fi
 
 
 def create_raw_data_ini(value_by_option_by_section = None):
-    """Creates raw_data.ini configureatio nfile
+    """Creates raw_data.ini configureation file
 
     Args:
       value_by_option_by_section(dict(dict)): Options value by section (Default value = None)
